added rollover and monitoring and optional parameters for duration and file size

+++ logger.go
@@ -10,45 +10,53 @@ import (
 	"time"
 )
 
-const prefix = "al_"
+const (
+	prefix         = "al_"
+	defaultMaxSize = 200 * 1024
+)
 
+var active = false
+var d time.Duration
+var maxsize int64
+var retentionDuration time.Duration
 var logfile *os.File
+var t *time.Ticker
 
-// Start changes output of log to be a file
+// Start the outpug log to file with default options
 func Start() error {
-	logdir := getApplicationDirectory()
-	if _, err := os.Stat(logdir); os.IsNotExist(err) {
-		if err := os.Mkdir(logdir, 0666); err != nil {
-			return err
-		}
-	}
+	return StartWithOptions(time.Hour, defaultMaxSize, 24*time.Hour)
+}
 
-	// getting the instance id
-	filename := os.Getenv("WEBSITE_INSTANCE_ID")
-	if len(filename) <= 6 {
-		filename = "local.log"
-	} else {
-		filename = fmt.Sprintf("%s_%s.log", filename[0:6], time.Now().Format(time.RFC3339))
-	}
+// StartWithOptions can be used to change default option
+func StartWithOptions(md time.Duration, mfs int64, rd time.Duration) error {
+	active = true
 
-	f, err := os.OpenFile(path.Join(logdir, prefix+filename), os.O_APPEND|os.O_CREATE|os.O_RDWR, 0666)
-	if err != nil {
-		return err
-	}
-
-	logfile = f
-
-	log.SetOutput(logfile)
+	d = md
+	maxsize = mfs
+	retentionDuration = rd
 
 	purgeFiles()
 
-	return nil
+	err := swapFile()
+	go monitor()
+
+	return err
 }
 
 // Stop closes the log file
 func Stop() {
+	active = false
+
+	log.SetOutput(os.Stderr)
+
+	if t != nil {
+		t.Stop()
+		t = nil
+	}
+
 	if logfile != nil {
 		logfile.Close()
+		logfile = nil
 	}
 }
 
@@ -63,13 +71,73 @@ func getApplicationDirectory() string {
 	return home
 }
 
+func swapFile() error {
+	logdir := getApplicationDirectory()
+	if _, err := os.Stat(logdir); os.IsNotExist(err) {
+		if err := os.Mkdir(logdir, 0666); err != nil {
+			return err
+		}
+	}
+
+	// getting the instance id
+	filename := os.Getenv("WEBSITE_INSTANCE_ID")
+	if len(filename) <= 6 {
+		filename = "local.log"
+	} else {
+		filename = fmt.Sprintf("%s_%s.log", filename[0:6], time.Now().Format("2006-Jan-2-15-04-05"))
+	}
+
+	f, err := os.Create(path.Join(logdir, prefix+filename))
+	if err != nil {
+		return err
+	}
+
+	logfile = f
+
+	log.SetOutput(logfile)
+
+	return nil
+}
+
+func rollover() {
+	if logfile == nil {
+		return
+	}
+
+	s, err := logfile.Stat()
+	if err != nil || s == nil {
+		Stop()
+	}
+
+	kb := s.Size() / 1024
+	if kb > maxsize {
+		log.SetOutput(os.Stderr)
+
+		logfile.Close()
+		logfile = nil
+
+		swapFile()
+	}
+}
+
 func purgeFiles() {
 	filepath.Walk(getApplicationDirectory(), func(p string, f os.FileInfo, err error) error {
-		if strings.HasPrefix(f.Name(), prefix) && strings.HasSuffix(f.Name(), ".log") {
-			if time.Since(f.ModTime()).Hours() > 24 {
+		if f != nil && strings.HasPrefix(f.Name(), prefix) && strings.HasSuffix(f.Name(), ".log") {
+			if time.Since(f.ModTime()) > retentionDuration {
 				os.Remove(p)
 			}
 		}
 		return nil
 	})
+}
+
+func monitor() {
+	t = time.NewTicker(d)
+	for {
+		<-t.C
+		if active {
+			rollover()
+		}
+		purgeFiles()
+	}
 }
